requirements.txt
Django==2.2.4
djangorestframework==3.10.2
auth0-python==3.9.1
